% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Caller.class.R
\name{call_target}
\alias{call_target}
\alias{get_call_target}
\title{Get call target from a Caller}
\usage{
call_target(x)

get_call_target(x)
}
\arguments{
\item{x}{an object of class "Caller"}
}
\value{
For \code{call_target}: a character object

For \code{get_call_target}: a function
}
\description{
call_target returns the name of the targetted function.

get_call_target returns the function itself.
}
