% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Caller.class.R
\name{print.Caller}
\alias{print.Caller}
\alias{$.Caller}
\alias{$<-.Caller}
\alias{[.Caller}
\alias{[<-.Caller}
\alias{[[.Caller}
\alias{[[<-.Caller}
\alias{names.Caller}
\alias{names<-.Caller}
\alias{with.Caller}
\alias{as.list.Caller}
\title{Methods for Caller object}
\usage{
\method{print}{Caller}(x, .hide_target = TRUE, ...)

\method{$}{Caller}(x, y)

\method{$}{Caller}(x, y) <- value

\method{[}{Caller}(x, y, ...)

\method{[}{Caller}(x, y, ...) <- value

\method{[[}{Caller}(x, y, ...)

\method{[[}{Caller}(x, y, ...) <- value

\method{names}{Caller}(x)

\method{names}{Caller}(x) <- value

\method{with}{Caller}(x, expr, ...)

\method{as.list}{Caller}(x, all.names = TRUE)
}
\arguments{
\item{all.names}{default to TRUE, whether to extract all names, see \link[base]{as.list.environment}}
}
\description{
Methods for Caller object
}
\details{
The \code{with} method provides a convenient way to expose the inner environment that Caller points to,
 without the need to de-callable it.
This is mainly used for the case where subsetting methods are sufficient
}
