% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-engine.R
\name{callme_engine}
\alias{callme_engine}
\title{An engine for knitr to handle C code compatible with \code{.Call()}}
\usage{
callme_engine(options)
}
\arguments{
\item{options}{the data passed in from the code chunk in Rmarkdown or
       Quarto etc.

\describe{
\item{compile}{Actually compile the code. default TRUE}
\item{headers}{automatically include minimal R headers in code. Default: TRUE}
\item{rcode}{Include R code in a code block beneath the C code}
}}
}
\value{
text block to be rendered in the document
}
\description{
An engine for knitr to handle C code compatible with \code{.Call()}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Set the engine in an initial chunk in the document
# Then use \code{callme} as the chunk engine
knitr::knit_engines$set(callme = callme::callme_engine)
\dontshow{\}) # examplesIf}
}
