% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{cflags_default}
\alias{cflags_default}
\title{Retrieve the default CFLAGS R uses to compile shared libraries}
\usage{
cflags_default()
}
\value{
character string of default CFLAGS for this R installation
}
\description{
Retrieve the default CFLAGS R uses to compile shared libraries
}
\examples{
cflags_default()
}
