% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Caller.class.R
\name{decallable}
\alias{decallable}
\title{De-callable a caller}
\usage{
decallable(x, ...)
}
\arguments{
\item{x}{a caller}
}
\value{
a de-callable object
}
\description{
This function decalls a Caller object, exposing what it hides
}
