% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_callable.R
\name{make_callable}
\alias{make_callable}
\title{Generate a callable subsettable object}
\usage{
make_callable(x, call_target = ".call", self)
}
\arguments{
\item{x}{an subsettable object}

\item{call_target}{the target binding for the caller; must be function.}
}
\value{
a "Caller" object
}
\description{
Function to create a callable object that wrap around x
}
