% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_callable.R
\name{set_caller_env}
\alias{set_caller_env}
\title{Helper methods to derive the suitable function}
\usage{
set_caller_env(x, fn_target)
}
\arguments{
\item{x}{the object}

\item{fn_target}{Target function. Must have in each method.}
}
\description{
Helper methods to derive the suitable function
}
