% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-engine.R
\name{callme_engine}
\alias{callme_engine}
\title{An engine for knitr to handle C code compatible with \code{.Call()}}
\usage{
callme_engine(options)
}
\arguments{
\item{options}{the data passed in from the code chunk in Rmarkdown or
Quarto etc.}
}
\value{
text block to be rendered in the document
}
\description{
An engine for knitr to handle C code compatible with \code{.Call()}
}
\examples{
\dontrun{
# Set the engine in an initial chunk in the document
# Then use \code{callme} as the chunk engine
knitr::knit_engines$set(callme = callme::callme_engine)
}

}
