% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_generator.R
\name{R6CallClass}
\alias{R6CallClass}
\title{R6Class; just callable}
\usage{
R6CallClass(
  classname = NULL,
  public = list(),
  private = NULL,
  active = NULL,
  inherit = NULL,
  lock_objects = TRUE,
  class = TRUE,
  portable = TRUE,
  lock_class = FALSE,
  cloneable = TRUE,
  callable_object_target = ".call",
  ...
)
}
\arguments{
\item{classname, public, private, active, inherit, lock_objects, class, portable, lock_class, cloneable, parent_env}{See \link[R6::R6Class]{R6::R6Class} for more details}

\item{callable_object_target}{if is NULL, R6 object created is not callable. Default will point to a public method name ".call"}

\item{...}{every thing else passed to R6::R6Class}
}
\description{
An extension to R6::R6Class, make it callable
}
